modularize into potato cam
